feat: refactor page and update layout and styling for consistency
