+++ app/src/components/Sidebar.tsx
@@ -33,10 +33,7 @@ export const Sidebar = ({
       {sidebarActions}
 
       {/* Prerequisites */}
-      <CollapsibleSection
-        title={<p className="ml-auto">Prerequisites</p>}
-        defaultOpen={true}
-      >
+      <CollapsibleSection title={<p className="ml-auto">Prerequisites</p>}>
         <ul className="space-y-2">
           {guide.prerequisites.map((prereq: GuideReference) => (
             <li
